Fix bitshift issue.

+++ engine/player/sc_player.cpp
@@ -5184,7 +5184,7 @@ action_priority_list_t* player_t::get_action_priority_list( const std::string& n
     a = new action_priority_list_t( name, this );
     a -> action_list_comment_str = comment;
     a -> internal_id = static_cast<int>( action_list_id_++ );
-    a -> internal_id_mask = ( 1 << a -> internal_id );
+    a -> internal_id_mask = ( 1 << static_cast<uint64_t>( a -> internal_id ) );
     if ( action_list_id_ == 64 )
     {
       sim -> errorf( "%s maximum number of action lists is 64", name_str.c_str() );
